add request interface.
