--- src/types/global.d.ts
@@ -1,14 +1,0 @@
-import "express";
-
-declare global {
-  namespace Express {
-    interface Request {
-      user?: {
-        id: string;
-        email: string;
-        role: string;
-        organizationId?: string;
-      };
-    }
-  }
-}
